\name{CompStratDesign-package}
\alias{CompStratDesign-package}
\alias{CompStratDesign}
\docType{package}
\title{
\packageTitle{CompStratDesign}
}
\description{
\packageDescription{CompStratDesign}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CompStratDesign}
\packageIndices{CompStratDesign}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CompStratDesign}

Maintainer: \packageMaintainer{CompStratDesign}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
